+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `Default`, `Debug`, `Clone`, `PartialEq`, and `Eq` for `Actions<C>` no longer impose trait bounds on `C`.
+
 ## [0.20.0] - 2025-11-15
 
 ### Added